MAINT: Move equation/variable groups to separate file
